git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@356 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/cstdkbe/memorystream.hpp
@@ -194,7 +194,7 @@ class MemoryStream
 
         MemoryStream &operator<<(COMPONENT_TYPE value)
         {
-            append<int8>(value);
+            append<int32>(value);
             return *this;
         }
 
@@ -300,7 +300,7 @@ class MemoryStream
         
         MemoryStream &operator>>(COMPONENT_TYPE &value)
         {
-            value = static_cast<COMPONENT_TYPE>(read<int8>());
+            value = static_cast<COMPONENT_TYPE>(read<int32>());
             return *this;
         }
